+++ changelog.md
@@ -4,6 +4,7 @@
 
 - structurizr-dsl: Adds support for element/relationship property expressions (https://github.com/structurizr/java/issues/297).
 - structurizr-dsl: Adds a way to specify the implied relationships strategy by a fully qualified class name via `!impliedRelationships`.
+- structurizr-dsl: Adds the ability to include single files as documentation (https://github.com/structurizr/java/issues/303). 
 
 ## 2.1.4 (18th June 2024)